+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+* `sources` string array on `Environment` to define .env files to be loaded
+
 <!--
 ### Added
 ### Changed